Refactoring SearchServiceImpl; moved all neo4j stuff to new Neo4jSearchServiceImpl class
