+++ corelib-search/src/main/java/eu/europeana/corelib/search/impl/SearchServiceImpl.java
@@ -27,12 +27,6 @@ import eu.europeana.corelib.edm.exceptions.MongoDBException;
 import eu.europeana.corelib.edm.exceptions.MongoRuntimeException;
 import eu.europeana.corelib.edm.exceptions.SolrTypeException;
 import eu.europeana.corelib.mongo.server.EdmMongoServer;
-import eu.europeana.corelib.neo4j.entity.CustomNode;
-import eu.europeana.corelib.neo4j.entity.Neo4jBean;
-import eu.europeana.corelib.neo4j.entity.Neo4jStructBean;
-import eu.europeana.corelib.neo4j.entity.Node2Neo4jBeanConverter;
-import eu.europeana.corelib.neo4j.exception.Neo4JException;
-import eu.europeana.corelib.neo4j.server.Neo4jServer;
 import eu.europeana.corelib.search.SearchService;
 import eu.europeana.corelib.search.model.ResultSet;
 import eu.europeana.corelib.search.query.MoreLikeThis;
@@ -76,7 +70,6 @@ import org.apache.solr.common.SolrException;
 import org.apache.solr.common.params.CursorMarkParams;
 import org.apache.solr.common.params.ModifiableSolrParams;
 import org.apache.solr.common.util.NamedList;
-import org.neo4j.graphdb.Node;
 import org.springframework.beans.factory.annotation.Value;
 
 import javax.annotation.Resource;
@@ -120,8 +113,6 @@ public class SearchServiceImpl implements SearchService {
     protected EdmMongoServer mongoServer;
     @Resource(name = "corelib_solr_mongoServer_id")
     protected EuropeanaIdMongoServer idServer;
-    @Resource(name = "corelib_solr_neo4jServer")
-    protected Neo4jServer neo4jServer;
 
     // provided by setSolrServer()
     private SolrServer solrServer;
@@ -749,41 +740,6 @@ public class SearchServiceImpl implements SearchService {
         }
     }
 
-    @Override
-    public List<Neo4jBean> getChildren(String rdfAbout, int offset, int limit) {
-        List<Neo4jBean> beans = new ArrayList<>();
-        long startIndex = offset;
-        List<CustomNode> children = neo4jServer.getChildren(rdfAbout, offset, limit);
-        for (CustomNode child : children) {
-            startIndex += 1L;
-            beans.add(Node2Neo4jBeanConverter.toNeo4jBean(child, startIndex));
-        }
-        return beans;
-    }
-
-    @Override
-    public List<Neo4jBean> getChildren(String rdfAbout, int offset) {
-        return getChildren(rdfAbout, offset, 10);
-    }
-
-    @Override
-    public List<Neo4jBean> getChildren(String rdfAbout) {
-        return getChildren(rdfAbout, 0, 10);
-    }
-
-    private Node getNode(String rdfAbout) throws Neo4JException {
-        return neo4jServer.getNode(rdfAbout);
-    }
-
-    @Override
-    public Neo4jBean getHierarchicalBean(String rdfAbout) throws Neo4JException {
-        Node node = getNode(rdfAbout);
-        if (node != null) {
-            return Node2Neo4jBeanConverter.toNeo4jBean(node, neo4jServer.getNodeIndex(node));
-        }
-        return null;
-    }
-
     private enum SuggestionTitle {
         TITLE("title", "Title"), DATE("when", "Time/Period"), PLACE("where",
                 "Place"), PERSON("who", "Creator"), SUBJECT("what", "Subject");
@@ -822,71 +778,6 @@ public class SearchServiceImpl implements SearchService {
         if (LOG.isDebugEnabled()) {
             LOG.debug(String.format("elapsed time (%s): %d", type, time));
         }
-    }
-
-    @Override
-    public List<Neo4jBean> getPrecedingSiblings(String rdfAbout, int offset, int limit) throws Neo4JException {
-        List<Neo4jBean> beans = new ArrayList<>();
-        List<CustomNode> precedingSiblings = neo4jServer.getPrecedingSiblings(rdfAbout, offset, limit);
-        long startIndex = neo4jServer.getNodeIndexByRdfAbout(rdfAbout) - offset;
-        for (CustomNode precedingSibling : precedingSiblings) {
-            startIndex -= 1L;
-            beans.add(Node2Neo4jBeanConverter.toNeo4jBean(precedingSibling, startIndex));
-        }
-        return beans;
-    }
-
-    @Override
-    public List<Neo4jBean> getPrecedingSiblings(String rdfAbout, int limit) throws Neo4JException {
-        return getPrecedingSiblings(rdfAbout, 0, 10);
-    }
-
-    @Override
-    public List<Neo4jBean> getPrecedingSiblings(String rdfAbout) throws Neo4JException {
-        return getPrecedingSiblings(rdfAbout, 10);
-    }
-
-    @Override
-    public List<Neo4jBean> getFollowingSiblings(String rdfAbout, int offset, int limit) throws Neo4JException {
-        List<Neo4jBean> beans = new ArrayList<>();
-        List<CustomNode> followingSiblings = neo4jServer.getFollowingSiblings(rdfAbout, offset, limit);
-        long startIndex = neo4jServer.getNodeIndexByRdfAbout(rdfAbout) + offset;
-        for (CustomNode followingSibling : followingSiblings) {
-            startIndex += 1L;
-            beans.add(Node2Neo4jBeanConverter.toNeo4jBean(followingSibling, startIndex));
-        }
-        return beans;
-    }
-
-    @Override
-    public List<Neo4jBean> getFollowingSiblings(String rdfAbout, int limit) throws Neo4JException {
-        return getFollowingSiblings(rdfAbout, 0, 10);
-    }
-
-    @Override
-    public List<Neo4jBean> getFollowingSiblings(String rdfAbout) throws Neo4JException {
-        return getFollowingSiblings(rdfAbout, 10);
-    }
-
-    @Override
-    public long getChildrenCount(String rdfAbout) throws Neo4JException {
-        return neo4jServer.getChildrenCount(getNode(rdfAbout));
-    }
-
-    // note that parents don't have their node indexes set in getInitialStruct
-    // because they have to be fetched separately; therefore this is done afterwards
-    @Override
-    public Neo4jStructBean getInitialStruct(String nodeId) throws Neo4JException {
-        return addParentNodeIndex(Node2Neo4jBeanConverter.toNeo4jStruct(neo4jServer.getInitialStruct(nodeId), neo4jServer.getNodeIndex(getNode(nodeId))));
-    }
-
-    private Neo4jStructBean addParentNodeIndex(Neo4jStructBean struct) throws Neo4JException {
-        if (!struct.getParents().isEmpty()) {
-            for (Neo4jBean parent : struct.getParents()) {
-                parent.setIndex(neo4jServer.getNodeIndex(getNode(parent.getId())));
-            }
-        }
-        return struct;
     }
 
     private static class PreEmptiveBasicAuthenticator implements HttpRequestInterceptor {
